Cleaning up and sorting functionality
